HDDS-12029. Move ozone debug recover to ozone admin om lease recover
